backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests>=2.32.0
urllib3>=2.0.0
chardet<6.0.0
charset-normalizer>=3.0.0,<4.0.0

psycopg2-binary
pydantic-settings
chromadb>=1.5.5
groq
google-genai
firecrawl-py
supabase
scrapy
langchain
langchain-community
langchain-groq
langchain-google-genai
pypdf
pandas
python-multipart
yfinance
tavily-python
gunicorn
celery
redis
beautifulsoup4
httpx
playwright
pdfplumber
camelot-py
sentence-transformers
jugaad-data